- Removed old file indexing code. - Switched file sizes from 32bits -> 64 bits so rs can handle large files. - combined Cache/Service Ids are now located in serialiser/rsserviceids.h
git-svn-id: http://svn.code.sf.net/p/retroshare/code/trunk@273 b45a01b8-16f6-495d-af2f-9b41ad6348cc

+++ libretroshare/src/dbase/cachestrapper.cc
@@ -639,7 +639,7 @@ bool CacheTransfer::RequestCache(CacheData &data, CacheStore *cbStore)
 
 
 /* to be overloaded */
-bool CacheTransfer::RequestCacheFile(RsPeerId id, std::string path, std::string hash, uint32_t size)
+bool CacheTransfer::RequestCacheFile(RsPeerId id, std::string path, std::string hash, uint64_t size)
 {
 	std::cerr << "CacheTransfer::RequestCacheFile() : from:" << id << " #";
 	std::cerr << hash << " size: " << size;
@@ -704,7 +704,7 @@ bool CacheTransfer::FailedCache(std::string hash)
 }
 
 
-bool    CacheTransfer::FindCacheFile(std::string hash, std::string &path, uint32_t &size)
+bool    CacheTransfer::FindCacheFile(std::string hash, std::string &path, uint64_t &size)
 {
 	CacheData data;
 	if (strapper->findCache(hash, data))
